fix: support sharkd path with spaces
process.spawn supports command and arg handling for the command parameter.
But it doesn't expect the path to be part of the command but to use the cwd option for that.
Use path.parse to split the sharkdPath into dir and base and pass these as args.
Added support for sharkdPath without dir as well defaulting to process.cwd().

Seems like the logic is not needed for tshark as there a seperator... is used.

Issue: about settings #29